en.text - removed harmful authentication advice
The original document suggested authenticating an OTR buddy through the OTR chat itself with questions that only the buddy would know. This is not actually sufficient (or any) protection against Man-In-The-Middle attacks, which are the entire reason that buddy authentication is so important.

Unfortunately, the true answer to secure buddy (really public key) authentication is rather more complicated for the user, but it must be done in such a way to ensure that the communication is actually secure.