en gpg-keys: restore gpg output; link to Debian + Ubuntu (#336 #339)
The original pgp output should be preserved as it trains readers to look for this layout.
Also the marker '^KEY-ID^' needs to be positioned at the correct place to make sense.

Debian and Ubuntu are preferred options to drive users away from windows.
They are responsive towards CVS and offer security patches when necessary.
Providing links to GNU and distrowatch for an informed decision.