clarify key splitting best practices
the encryption subkey recommendation is implemented by default in Debian jessie and stretch, as far as I can tell.

furthermore, the guide was telling us to keep primary keys offline, but didn't tell us how to do that, or how to create new subkeys. add some information about how to do that.