link to the 2.1 review guide
it will be useful for people to know what version of GnuPG was the guide written for. they will also find useful tips and quick fixes on github that way, and that will encourage people to participate in writing the guide.

i added an edit link because it's difficult to find the actual page from the "edit" button at the bottom of the page, which only brings you to the top of the repo, and not directly on the page.