sync update-resolv-conf script with Debian
The script apparently comes from Debian, but it looks really, really
old. There were changes done to the script since then including
improvements on error handling and refactoring that make the script
easier to read and audit.

The script content was synchronized with the OpenVPN 2.4.6-1 Debian
package, as available here:

https://sources.debian.org/data/main/o/openvpn/2.4.6-1/debian/update-resolv-conf