Promote riseup-vpn to be the main VPN provided.
The entire structure of our help pages were oriented around 'red' and 'black'
services, where 'black' services were those that required the Bitmask
application to be used and a separate account to be created.

In order to de-emphasize Bitmask, as it is being deprecated, and promote
RiseupVPN as the preferred solution, some changes to how things are organized
around the "Red" vs "Black" structure.