Update en.md
Lubuntu doesn't come with python, which needs to be installed. Perhaps python-minimal is enough? Feel free to rephrase the line, as English is not my native language.

The error produced is: Run install hook of snap if present (run hook "install": /usr/bin/env: "python": The file or folder does not exist)