riseup-vpn doesn't leave the firewall either the pid file after a kill
Remove the documentation about turning off the firewall and removing the
PID file. Since more than a year riseup-vpn does handle signals and do
the cleaning properly.

Also this instructions are confusing if you are using the debian package
instead of the snap, as the location of bitmask-root is a different one.