fix tbird use_status_for_biff setting
We offer this as a way to reduce overhead/server load, but it seems to
be the wrong setting. Everywhere I look including [MozillaZine][1] shows
the setting being prepended with `mail.imap`, which indeed exists as a
default setting and is consistent with the hierarchical formatting of
the other settings.

[1]: http://kb.mozillazine.org/Mail_and_news_settings