Fix tense and remove ambiguity about Bitmask
The wording "it had" left one to wonder what thing *had* features that made Bitmask great and why that would be relevant now.  Presumably, this was intended to say that RiseupVPN, the replacement for Riseup Black, which was based on Bitmask, *has* all of the features that made Bitmask great.