Clean up old and dangerous apt info
anarcat added a warning one year ago. Even if people are still using
bullseye, this advice is wrong/dangerous. If a bullseye-backport becomes
available we can add that.

Removed from the translations that had it, a couple lack the whole
"Package Installation" section entirely, I did not attempt to add it.

It still needs to be removed upstream
https://0xacab.org/leap/bitmask-vpn/-/issues/693