Update Web Browsing Exensions en.text
Corrected uBlock Origin - it is no longer available to Chrome users. Chrome users need to use uBlock Origin Lite.

Removed HTTPS Everywhere - This is a pretty standard browser feature now. Default Firefox has it.

Removed Privacy Badger - Privacy Badger's dynamic tracker blocking is no longer used, as it was found to be a fingerprinting vector. It is still maintained, but with a static block list, it can easily be replaced by uBlock Origin.

Added Port Authority - Port Authority is a low setup and 0 maintenance extension that blocks port scans, scans which can view your private network. It is a less common or useful invasion of privacy, but there are websites that do it. The extension is open source, of course, but currently limited to Firefox-based browsers.